chore: move auto implemented KvApi methods to Ext trait
