+++ src/meta/control/src/lua_support.rs
@@ -21,7 +21,7 @@ use databend_common_base::base::BuildInfoRef;
 use databend_common_meta_client::errors::CreationError;
 use databend_common_meta_client::ClientHandle;
 use databend_common_meta_client::MetaGrpcClient;
-use databend_common_meta_kvapi::kvapi::KVApi;
+use databend_common_meta_kvapi::kvapi::KvApiExt;
 use databend_common_meta_types::UpsertKV;
 use mlua::Lua;
 use mlua::LuaSerdeExt;
