+++ src/meta/kvapi/src/kvapi/mod.rs
@@ -20,6 +20,7 @@ mod key;
 mod key_builder;
 mod key_codec;
 mod key_parser;
+mod kv_api_ext;
 mod message;
 mod pair;
 mod prefix;
@@ -40,6 +41,7 @@ pub use key::KeyError;
 pub use key_builder::KeyBuilder;
 pub use key_codec::KeyCodec;
 pub use key_parser::KeyParser;
+pub use kv_api_ext::KvApiExt;
 pub use message::GetKVReply;
 pub use message::GetKVReq;
 pub use message::ListKVReply;
