+++ src/meta/service/src/meta_service/meta_leader.rs
@@ -19,6 +19,7 @@ use std::time::SystemTime;
 use anyerror::AnyError;
 use databend_common_meta_client::MetaGrpcReadReq;
 use databend_common_meta_kvapi::kvapi::KVApi;
+use databend_common_meta_kvapi::kvapi::KvApiExt;
 use databend_common_meta_sled_store::openraft::ChangeMembers;
 use databend_common_meta_stoerr::MetaStorageError;
 use databend_common_meta_types::node::Node;
