+++ src/meta/service/tests/it/grpc/metasrv_connection_error.rs
@@ -24,6 +24,7 @@ use databend_common_meta_client::errors::CreationError;
 use databend_common_meta_client::ClientHandle;
 use databend_common_meta_client::MetaGrpcClient;
 use databend_common_meta_kvapi::kvapi::KVApi;
+use databend_common_meta_kvapi::kvapi::KvApiExt;
 use databend_common_meta_types::UpsertKV;
 use log::info;
 use test_harness::test;
