+++ src/meta/service/tests/it/grpc/metasrv_grpc_api.rs
@@ -18,6 +18,7 @@ use std::collections::HashSet;
 use databend_common_base::base::tokio;
 use databend_common_base::base::Stoppable;
 use databend_common_meta_kvapi::kvapi::KVApi;
+use databend_common_meta_kvapi::kvapi::KvApiExt;
 use databend_common_meta_kvapi::kvapi::UpsertKVReply;
 use databend_common_meta_types::SeqV;
 use databend_common_meta_types::UpsertKV;
