+++ src/meta/service/tests/it/grpc/metasrv_grpc_tls.rs
@@ -17,7 +17,7 @@ use std::time::Duration;
 use databend_common_grpc::RpcClientTlsConfig;
 use databend_common_meta_api::TableApi;
 use databend_common_meta_client::MetaGrpcClient;
-use databend_common_meta_kvapi::kvapi::KVApi;
+use databend_common_meta_kvapi::kvapi::KvApiExt;
 use databend_common_meta_types::MetaClientError;
 use databend_common_meta_types::MetaError;
 use databend_common_meta_types::MetaNetworkError;
