+++ src/meta/service/tests/it/meta_node/meta_node_kv_api_expire.rs
@@ -15,7 +15,7 @@
 use std::time::Duration;
 
 use databend_common_base::base::tokio::time::sleep;
-use databend_common_meta_kvapi::kvapi::KVApi;
+use databend_common_meta_kvapi::kvapi::KvApiExt;
 use databend_common_meta_types::Cmd;
 use databend_common_meta_types::LogEntry;
 use databend_common_meta_types::MatchSeq;
