+++ src/query/service/src/interpreters/common/util.rs
@@ -16,7 +16,7 @@ use std::sync::Arc;
 
 use databend_common_catalog::table_context::TableContext;
 use databend_common_exception::Result;
-use databend_common_meta_kvapi::kvapi::KVApi;
+use databend_common_meta_kvapi::kvapi::KvApiExt;
 use databend_common_users::UserApiProvider;
 
 /// Checks if a duplicate label exists in the meta store.
