+++ src/query/service/src/task/meta.rs
@@ -19,6 +19,7 @@ use databend_common_base::runtime::block_on;
 use databend_common_exception::Result;
 use databend_common_meta_client::ClientHandle;
 use databend_common_meta_kvapi::kvapi::KVApi;
+use databend_common_meta_kvapi::kvapi::KvApiExt;
 use databend_common_meta_semaphore::acquirer::Permit;
 use databend_common_meta_semaphore::Semaphore;
 use databend_common_meta_types::MatchSeq;
