+++ src/query/storages/basic/src/result_cache/meta_manager.rs
@@ -17,6 +17,7 @@ use std::time::Duration;
 
 use databend_common_exception::Result;
 use databend_common_meta_kvapi::kvapi::KVApi;
+use databend_common_meta_kvapi::kvapi::KvApiExt;
 use databend_common_meta_store::MetaStore;
 use databend_common_meta_types::MatchSeq;
 use databend_common_meta_types::MetaSpec;
